first-pass at splitting up userspace
right now, it's all backwards. What _should_ happen is that you call
`cargo xbuild` on a specific userspace (e.g: k1, k2, etc...), where that
crate includes the choochoos kernel (i.e: runtime) as a dependency.

The only reason I didn't do this is because I have a sneaking suspicion
that'll make the build process real ugly.

Quite frankly, what _really_ aught to happen is that each individual
userspace is compiled to it's own staticlib, with it's own panic handler
(and allocator?) which uses syscalls to do most things.