nameserver impl + syscall improvements + mixed rust/c userspace
actually implementing the nameserver exposed out lots of misc bugs
and points of improvement in the Rust userspace library. notably, the
syscall implementation is a lot more robust now, and leverages the abi
crate as the source of truth for syscall numbers and function
signatures, preventing transcription errors that cause undefined
behavior.

additionally, i've sprinkled the `never`/`!` type throughout the syscall
function signatures, which brings with it some cool optimization
opportunties and correctness enhancements. I've even included a brief
writeup about this in the README.md

lastly, I've validated the nameserver implementation against my existing
C userspace code, which required adding a "stub" distro that links in
the Rust userspace code with existing C code. This is different from
using the DISTRO env var directly, as that requires the linked binary to
include it's own nameserver and syscall implementations.